// src/types.ts
export interface UserInterface {
    id: string;
    initials: string;
    name: string;
    age: number;
    gender: string;
    date: string;
    sessionTime: string;
    type: string;
    status: string;
    email?: string;
    telefone?: string;
    whatsapp?: boolean;
    cpf?: string;
    rg?: string;
  }